[iris] Isolate Kubernetes cache maintenance
Move cache reclamation into a dedicated node-agent module and report sweep failures through telemetry. Run the DaemonSet in cache-only clusters and share the default task-cache path across Kubernetes components.
