+++ lib/iris/src/iris/cluster/platforms/k8s/controller.py
@@ -33,7 +33,11 @@ from iris.cluster.config import (
 from iris.cluster.endpoints import LOG_SERVER_ENDPOINT_NAME
 from iris.cluster.inject_env import TASK_ENV_SECRET_NAME, collect_inject_env, projects_task_env_secret
 from iris.cluster.node_agent import SERVICE_NAME as _NODE_AGENT_NAME
-from iris.cluster.platforms.k8s.constants import COREWEAVE_INTERRUPTABLE_TOLERATION, NVIDIA_GPU_TOLERATION
+from iris.cluster.platforms.k8s.constants import (
+    COREWEAVE_INTERRUPTABLE_TOLERATION,
+    DEFAULT_TASK_CACHE_DIR,
+    NVIDIA_GPU_TOLERATION,
+)
 from iris.cluster.platforms.k8s.kueue_manifests import (
     IRIS_WORKLOAD_PRIORITY_CLASSES,
     build_workload_priority_class,
@@ -512,9 +516,13 @@ class K8sControllerProvider:
 
         self.ensure_kueue_queues(config)
         self.ensure_priority_classes()
-        if config.finelog.config or LOG_SERVER_ENDPOINT_NAME in config.endpoints:
+        if (
+            config.finelog.config
+            or LOG_SERVER_ENDPOINT_NAME in config.endpoints
+            or config.kubernetes_provider.cache_max_age is not None
+        ):
             cache_dir = (
-                config.kubernetes_provider.cache_dir or "/cache"
+                config.kubernetes_provider.cache_dir or DEFAULT_TASK_CACHE_DIR
                 if config.kubernetes_provider.cache_max_age is not None
                 else None
             )
